added a new side bar
Basically, the sidebar has two tabs, contacts and conversations, each of
these tabs will show the user either their contacts list or
conversations list. Anyways there is also a button on the bottom of the
tab that you will use to add contacts or add conversations. That is
basically it for the sidebar